[CONJ-1009] improve performance reading result-set of big size part 3
 * JSON metadata correction

+++ src/main/java/org/mariadb/jdbc/client/column/StringColumn.java
@@ -398,10 +398,10 @@ public class StringColumn extends ColumnDefinitionPacket implements ColumnDecode
       Timestamp timestamp;
       synchronized (cal) {
         cal.clear();
-        cal.set(year, month - 1, (int) dayOfMonth, hour, minutes, seconds);
+        cal.set(year, month - 1, dayOfMonth, hour, minutes, seconds);
         timestamp = new Timestamp(cal.getTimeInMillis());
       }
-      timestamp.setNanos((int) (microseconds * 1000));
+      timestamp.setNanos(microseconds * 1000);
       return timestamp;
 
     } catch (DateTimeException dte) {
